2010-04-25	Joel Sherrill <joel.sherrilL@OARcorp.com>
	* console/uart.c: Fix warnings.

+++ c/src/lib/libbsp/arm/gp32/console/uart.c
@@ -40,7 +40,7 @@ int dbg_dly;
 static int     uart_first_open(int major, int minor, void *arg);
 static int     uart_last_close(int major, int minor, void *arg);
 static int     uart_read(int minor);
-static int     uart_write(int minor, const char *buf, int len);
+static ssize_t uart_write(int minor, const char *buf, size_t len);
 static void    uart_init(int minor);
 static void    uart_write_polled(int minor, char c);
 static int     uart_set_attributes(int minor, const struct termios *t);
@@ -153,7 +153,7 @@ static int uart_read(int minor)
  *
  * return 1 on success, -1 on error
  */
-static int uart_write(int minor, const char *buf, int len)
+static ssize_t uart_write(int minor, const char *buf, size_t len)
 {
     int i;
